test(#192): fold the pool flatness sweep into the registry's own test
A test file with no corresponding source module breaks co-location; the
sweep is one case beside the registry it reads, computation inlined.
